simplify logging
Use the module logger to log all messages from yacc and the parser. Explicitly pass a dedicated log dumped to a file in `rebuild_table`.
Do not pass an `errorlog` to avoid masking warnings and other messages during construction of the parser.
Add an attribute `logger` to easily substitute the parser's logger.