add semantic `else` and make default
because the Promela language reference defines `else` semantically, with respect to the program graph, and not the source listing.

A semantic `else` escapes statements that have the same source node in the program graph. For each program graph node, at most one outgoing edge can be labeled with an `else`.

A syntactic `else` escapes only statements in the same syntactic "options" construct. Multiple `else` statements can label edges with the same source node in the program graph. The syntactically innermost `else` takes precedence. The rest block, because one of their alternative options is always executable (due to the inner `else`).